Move deletion of old GTID rows to slave background thread
Intermediate commit, incomplete patch.

In record_gtid(), we no longer do any deletion of old rows in
mysql.gtid_slave_pos. Instead, in update(), we increment a counter,
and when it reaches some fixed value, we signal the slave background
thread.

The background thread then deletes all pending old rows in a batch.

This simplifies the logic for deleting old rows. It also removes the
cost of deletions from the replication of each transaction, removing
close to half the contention on LOCK_slave_state. It adds the cost of
an extra transaction per batch deletion (per mysql.gtid_slave_posXXX
table), but this can be reduced by increasing the batch size.